Change PIN ViewModel two steps
