+++ ni_sdk/src/main/java/ae/network/nicardmanagementsdk/presentation/ui/change_pin/ChangePinViewModel.kt
@@ -4,51 +4,38 @@ import ae.network.nicardmanagementsdk.R
 import ae.network.nicardmanagementsdk.core.IChangePinCore
 import ae.network.nicardmanagementsdk.network.utils.ConnectionModel
 import ae.network.nicardmanagementsdk.network.utils.IConnection
-import ae.network.nicardmanagementsdk.presentation.ui.set_pin.SetPinViewModelBase
+import ae.network.nicardmanagementsdk.presentation.ui.set_pin.SetPinTwoStepViewModelBase
 import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.viewModelScope
-import kotlinx.coroutines.launch
 
-class ChangePinViewModel (
+class ChangePinViewModel(
     private val changePinCoreComponent: IChangePinCore,
-    private val connectionLiveData: IConnection<ConnectionModel>
-) : SetPinViewModelBase() {
+    connectionLiveData: IConnection<ConnectionModel>
+) : SetPinTwoStepViewModelBase(connectionLiveData) {
 
     override val navTitle = MutableLiveData(R.string.change_pin_title)
     override val screenTitle = MutableLiveData(R.string.change_pin_description_enter_current_pin)
-    private var oldPin = ""
-    private var newPin = ""
+
+    private var currentPin = ""
     private var isCurrentPinSetup = true
 
 
     override fun onDoneImageButtonTap() {
         if (isCurrentPinSetup) {
-            oldPin = inputString
-            resetState()
-            screenTitle.value = R.string.change_pin_description_enter_new_pin
-            isCurrentPinSetup = false
+            captureCurrentPin()
         } else {
-            newPin = inputString
-            viewModelScope.launch {
-                if (connectionLiveData.hasInternetConnectivity) {
-                    isVisibleProgressBar.value = true
-                    val result = changePinCoreComponent.makeNetworkRequest(
-                        oldPin,
-                        newPin
-                    )
-                    isVisibleProgressBar.value = false
-                    onResultSingleLiveEvent.value = result
-                }
-            }
+            setPinTwoSteps(networkRequest = {
+                changePinCoreComponent.makeNetworkRequest(
+                    currentPin,
+                    secondTimePinValue
+                )
+            })
         }
     }
 
-    fun resetChangePinState() {
-        oldPin = ""
-        newPin = ""
-        isCurrentPinSetup = true
-        screenTitle.value = R.string.change_pin_description_enter_current_pin
+    private fun captureCurrentPin() {
+        currentPin = inputString
         resetState()
+        screenTitle.value = R.string.change_pin_description_enter_new_pin
+        isCurrentPinSetup = false
     }
-
 }
